Moved testing utils outside of ml_worker module
