+++ giskard/push/push_test_catalog/catalog.py
@@ -20,11 +20,11 @@ from giskard.ml_worker.testing.functions.transformation import mad_transformatio
 from giskard.ml_worker.testing.registry.decorators import test
 from giskard.ml_worker.testing.registry.slicing_function import SlicingFunction
 from giskard.ml_worker.testing.test_result import TestResult
-from giskard.ml_worker.testing.utils import Direction, check_slice_not_empty
 from giskard.models.base import BaseModel
 from giskard.testing.tests.calibration import test_overconfidence_rate, test_underconfidence_rate
 from giskard.testing.tests.metamorphic import test_metamorphic_invariance
 from giskard.testing.tests.performance import _test_diff_prediction, test_f1, test_rmse
+from giskard.testing.utils.utils import Direction, check_slice_not_empty
 
 
 @test(name="Slicing RMSE difference", tags=["performance", "regression", "ground_truth", "push"])
