+++ giskard/testing/tests/drift.py
@@ -1,6 +1,3 @@
-import typing
-from typing import List, Optional
-
 import inspect
 import numbers
 import re
@@ -9,16 +6,17 @@ from collections import Counter
 
 import numpy as np
 import pandas as pd
+import typing
 from scipy.stats import chi2, ks_2samp
 from scipy.stats.stats import Ks_2sampResult, wasserstein_distance
+from typing import List, Optional
 
 from giskard.datasets.base import Dataset
 from giskard.ml_worker.testing.registry.decorators import test
 from giskard.ml_worker.testing.registry.slicing_function import SlicingFunction
 from giskard.ml_worker.testing.test_result import TestMessage, TestMessageLevel, TestResult
-from giskard.ml_worker.testing.utils import check_slice_not_empty, validate_classification_label
 from giskard.models.base import BaseModel
-
+from giskard.testing.utils.utils import check_slice_not_empty, validate_classification_label
 from . import debug_description_prefix
 
 other_modalities_pattern = "^other_modalities_[a-z0-9]{32}$"
