+++ giskard/testing/tests/metamorphic.py
@@ -1,7 +1,6 @@
-from typing import Optional
-
 import numpy as np
 import pandas as pd
+from typing import Optional
 
 from giskard.core.core import SupportedModelTypes
 from giskard.datasets.base import Dataset
@@ -9,22 +8,21 @@ from giskard.llm import LLMImportError
 from giskard.ml_worker.testing.registry.decorators import test
 from giskard.ml_worker.testing.registry.slicing_function import SlicingFunction
 from giskard.ml_worker.testing.registry.transformation_function import TransformationFunction
-from giskard.ml_worker.testing.stat_utils import (
+from giskard.ml_worker.testing.test_result import TestMessage, TestMessageLevel, TestResult
+from giskard.ml_worker.utils.logging import timer
+from giskard.models.base import BaseModel
+from giskard.models.utils import fix_seed
+from giskard.testing.utils.stat_utils import (
     equivalence_t_test,
     equivalence_wilcoxon,
     paired_t_test,
     paired_wilcoxon,
 )
-from giskard.ml_worker.testing.test_result import TestMessage, TestMessageLevel, TestResult
-from giskard.ml_worker.testing.utils import (
+from giskard.testing.utils.utils import (
     Direction,
     check_slice_not_empty,
     validate_classification_label,
 )
-from giskard.ml_worker.utils.logging import timer
-from giskard.models.base import BaseModel
-from giskard.models.utils import fix_seed
-
 from . import debug_description_prefix
