+++ giskard/testing/tests/performance.py
@@ -1,6 +1,4 @@
 """Performance tests"""
-from typing import Optional
-
 import numpy as np
 from sklearn.metrics import (
     accuracy_score,
@@ -12,19 +10,19 @@ from sklearn.metrics import (
     recall_score,
     roc_auc_score,
 )
+from typing import Optional
 
 from giskard.datasets.base import Dataset
 from giskard.ml_worker.testing.registry.decorators import test
 from giskard.ml_worker.testing.registry.slicing_function import SlicingFunction
 from giskard.ml_worker.testing.test_result import TestResult
-from giskard.ml_worker.testing.utils import Direction, check_slice_not_empty
 from giskard.models.base import BaseModel
 from giskard.models.utils import np_type_to_native
 from giskard.testing.tests.debug_slicing_functions import (
     incorrect_rows_slicing_fn,
     nlargest_abs_err_rows_slicing_fn,
 )
-
+from giskard.testing.utils.utils import Direction, check_slice_not_empty
 from . import debug_description_prefix
