+++ giskard/testing/tests/statistic.py
@@ -1,17 +1,15 @@
 """Statistical tests"""
-from typing import Optional
-
 import numbers
 
 import numpy as np
+from typing import Optional
 
 from giskard.datasets.base import Dataset
 from giskard.ml_worker.testing.registry.decorators import test
 from giskard.ml_worker.testing.registry.slicing_function import SlicingFunction
 from giskard.ml_worker.testing.test_result import TestMessage, TestMessageLevel, TestResult
-from giskard.ml_worker.testing.utils import check_slice_not_empty, validate_classification_label
 from giskard.models.base import BaseModel
-
+from giskard.testing.utils.utils import check_slice_not_empty, validate_classification_label
 from . import debug_description_prefix
