+++ tests/test_performance.py
@@ -1,9 +1,11 @@
 import re
-import pytest
+
 import pandas as pd
+import pytest
+
 import giskard.testing.tests.performance as performance
-from giskard.ml_worker.testing.utils import Direction
 from giskard.ml_worker.testing.registry.slicing_function import SlicingFunction, slicing_function
+from giskard.testing.utils.utils import Direction
 
 
 @pytest.mark.parametrize(
